+++ lib/src/main/java/de/kuschku/libquassel/session/SocketAddress.kt
@@ -2,6 +2,6 @@ package de.kuschku.libquassel.session
 
 import java.net.InetSocketAddress
 
-data class SocketAddress(val host: String, val port: Short) {
-  fun data() = InetSocketAddress(host, port.toInt())
+data class SocketAddress(val host: String, val port: Int) {
+  fun data() = InetSocketAddress(host, port)
 }
